write-behind/write-pubsub/requirements.txt
google-cloud-pubsub==2.13.9
uwsgi==2.0.21
flask==2.1.0
